Added modules to our project
